feat(Cohort): add dataset attribute to Cohort target
This PR will solve the raised issue #75 where the Cohort target
currently does not have access to its own dataset (as an attribute) in
order to save output files. The proposed solution is as follows.

We keep the analysis_dataset property available in order to maintain the
current functionality of existing pipelines.

This also means in future this gives the power to the Pipeline
Developers to choose the most appropriate place to save the outputs:

* In the analysis_dataset as defined in the config provided to the
  analysis_runner (the same one as used in the MultiCohortStage)

* The dataset as pulled from the metadata in Metamist to save each
  respective Cohorts results in its respective dataset location.

This naming convention is also consistent with say SequencingGroups
which don't have an analysis_dataset attribute only a dataset attribute
which is determined entirely by metadata on that individual sequencing
group (pulled from metamist).

We could discuss whether it would be a helpful feature for all Target
types to have access to this analysis_dataset property or not. That
would come down to whether it's needed, or useful or would just cause
confusion.

SET-490, Resolves #75