ci(test): sonarqube scan every pr and create it's own project
This is the (slightly hacky) work around for being limited to the
community version where we can't do separate scans on each PR. The
solution is make a new sonarqube project for every project-pr-<pr-name>
and remove that project on merge to main (where a scan on main will
then be performed). A little gross, and the main downside is creating
an unnecessary number of projects cluttering the ui and also increasing
the required specs on our sonarqube instance.