fix(setuptools): add where syntax to find all sub-folders
The old build command was not finding any of the sub folders
(show_workflow and targets) in the src folder. In removing the use
of setuptools scm (which we didn't need since semantic release
handles versioning) and switching back to the simple setuptools
for the build setting the package-dir to "src" was insufficient.
Now with the find syntax added it finds all the files and sub folders
in the "src" folder.

SET-653